extracted interface and moved the ResourceManager to core
