+++ core/src/main/java/org/polypheny/db/filemanagement/FileSystemManagerImpl.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.polypheny.db.catalog;
+package org.polypheny.db.filemanagement;
 
 
 import java.io.File;
@@ -26,22 +26,22 @@ import java.util.List;
 /**
  * The FileSystemManager should handle all dynamically needed resources and sort them
  */
-public class FileSystemManager {
+public class FileSystemManagerImpl extends FileSystemManager {
 
     static FileSystemManager fileSystemManager = null;
-    final File root = new File(System.getProperty("user.home"), ".polypheny");
+    final File root = new File( System.getProperty( "user.home" ), ".polypheny" );
     final List<File> dirs = new ArrayList<>();
 
 
     public static FileSystemManager getInstance() {
         if ( fileSystemManager == null ) {
-            fileSystemManager = new FileSystemManager();
+            fileSystemManager = new FileSystemManagerImpl();
         }
         return fileSystemManager;
     }
 
 
-    private FileSystemManager() {
+    private FileSystemManagerImpl() {
         if ( !root.exists() ) {
             root.mkdir();
         }
@@ -50,9 +50,11 @@ public class FileSystemManager {
 
     /**
      * Registers a new data folder
+     *
      * @param path the path of the new folder
      * @return the file object for the directory
      */
+    @Override
     public File registerDataFolder( String path ) {
         File file = new File( this.root, path );
         if ( !file.exists() ) {
@@ -63,17 +65,21 @@ public class FileSystemManager {
         return file;
     }
 
-    public File registerNewFile( String pathToFile ){
+
+    @Override
+    public File registerNewFile( String pathToFile ) {
         return registerNewFile( root, pathToFile );
     }
 
 
     /**
      * places a new file in a specific path, if no path is specified, it is places in the root path
+     *
      * @param path path to the folder in which the file should be placed
      * @param pathToFile the file and its parent paths
      * @return the file object of the file itself
      */
+    @Override
     public File registerNewFile( File path, String pathToFile ) {
         File file = new File( path, pathToFile
                 .replace( "//", "/" )
